Refactors the App.jsx
* Segregates the post form elements to its own components

+++ frontend/src/App.tsx
@@ -1,5 +1,11 @@
-import { useEffect, useReducer, useState } from "react";
+import { type SubmitEvent, useEffect, useReducer, useState } from "react";
 import { type Action, type Post, Reducer } from "./reducer.ts";
+
+type Dispatch = (action: Action) => void;
+
+type id = string;
+
+type Body = { title: string; body: string } | { id: string };
 
 type Feed = {
   post: Post;
@@ -22,10 +28,7 @@ type FeedProps = {
   deleteThePost: (action: Action) => void;
 };
 
-const fetchPost = (
-  endPoint: string,
-  body: { title: string; body: string } | { id: string },
-) =>
+const fetchPost = (endPoint: string, body: Body) =>
   fetch(`http://localhost:8000/${endPoint}`, {
     method: "post",
     body: JSON.stringify(body),
@@ -35,6 +38,46 @@ const fetchPost = (
 
 const FormTitle = () => <h1>Create Post</h1>;
 
+const PostDescription = ({ setBody }) => (
+  <>
+    <p>Body</p>
+    <textarea
+      name="body"
+      placeholder="Write your post..."
+      rows={5}
+      cols={50}
+      onChange={(e) => setBody(e.target.value)}
+    >
+    </textarea>
+  </>
+);
+
+const PostTitle = ({ setTitle }) => (
+  <>
+    <p>Title</p>
+    <input
+      type="text"
+      placeholder="Enter a title..."
+      name="title"
+      onChange={(e) => setTitle(e.target.value)}
+      required
+    />
+  </>
+);
+
+const AddPost = async (
+  e: SubmitEvent<HTMLFormElement>,
+  title: string,
+  desc: string,
+  saveThePost: Dispatch,
+) => {
+  e.preventDefault();
+  const body = { title, body: desc };
+
+  const _id: id = await fetchPost("add-post", body);
+  saveThePost({ act: "add-post", title, _id, body: desc });
+};
+
 const DisplayForm = ({ saveThePost }: DisplayFormProps) => {
   const [title, setTitle] = useState("");
   const [body, setBody] = useState("");
@@ -42,64 +85,62 @@ const DisplayForm = ({ saveThePost }: DisplayFormProps) => {
   return (
     <form
       onSubmit={async (e) => {
-        e.preventDefault();
-        const _id: string = await fetchPost("add-post", { title, body });
-        saveThePost({ act: "add-post", title, _id, body });
+        await AddPost(e, title, body, saveThePost);
       }}
     >
       <FormTitle />
-      <p>Title</p>
-      <input
-        type="text"
-        placeholder="Enter a title..."
-        name="title"
-        onChange={(e) => setTitle(e.target.value)}
-        required
-      />
-      <p>Body</p>
-      <textarea
-        name="body"
-        placeholder="Write your post..."
-        rows={5}
-        cols={50}
-        onChange={(e) => setBody(e.target.value)}
-      >
-      </textarea>
+      <PostTitle setTitle={setTitle} />
+      <PostDescription setBody={setBody} />
       <button type="submit">Post</button>
     </form>
   );
 };
+
+const confirmTheDelete = async (post: Post, deleteThePost: Dispatch) => {
+  const confirmMsg = `Do you want to delete post: ${post.title}`;
+  const body = { id: post._id };
+
+  if (confirm(confirmMsg)) {
+    await fetchPost("delete-post", body);
+    deleteThePost({ act: "delete-post", post });
+  }
+};
+
+const DeleteButton = ({ post, deleteThePost }) => (
+  <>
+    <button
+      type="button"
+      onClick={async () => {
+        await confirmTheDelete(post, deleteThePost);
+      }}
+    >
+      Delete
+    </button>
+    <hr />
+  </>
+);
 
 const Feed = ({ post, deleteThePost }: FeedProps) => {
   return (
     <>
       <h2>{post.title}</h2>
       <p>{post.body}</p>
-      <button
-        type="button"
-        onClick={async () => {
-          await fetchPost("delete-post", { id: post._id });
-          deleteThePost({ act: "delete-post", post });
-        }}
-      >
-        Delete
-      </button>
-      <hr />
+      <DeleteButton post={post} deleteThePost={deleteThePost} />
     </>
   );
 };
 
 const CreateFeed = ({ posts, deleteThePost }: DeletePost) =>
-  posts.map((p) => <Feed post={p} key={p._id} deleteThePost={deleteThePost} />);
+  posts.map((post) => (
+    <Feed post={post} key={post._id} deleteThePost={deleteThePost} />
+  ));
 
-const DisplayFeed = ({ posts, deleteThePost }: DeletePost) => {
-  return (
-    <div>
-      <h1>Feed</h1>
-      <CreateFeed posts={posts} deleteThePost={deleteThePost} />
-    </div>
-  );
-};
+const DisplayFeed = ({ posts, deleteThePost }: DeletePost) => (
+  <div>
+    <h1>Feed</h1>
+    <CreateFeed posts={posts} deleteThePost={deleteThePost} />
+  </div>
+);
 
 const App = () => {
   const [posts, dispatch] = useReducer(Reducer, []);
